move configuration classes to more specific packages
