+++ eva-accession-release/src/main/java/uk/ac/ebi/eva/accession/release/configuration/readers/ContigReaderConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.release.configuration;
+package uk.ac.ebi.eva.accession.release.configuration.readers;
 
 import com.mongodb.MongoClient;
 import org.slf4j.Logger;
