+++ eva-accession-release/src/main/java/uk/ac/ebi/eva/accession/release/configuration/writers/ContigWriterConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.release.configuration;
+package uk.ac.ebi.eva.accession.release.configuration.writers;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
